fix: resolve import conflict between utils.py and utils/ directory
Move CMS versioning helpers from utils.py to utils/__init__.py to fix
ImportError. Python's import system treats the directory as a package
and ignores the file, causing the `has_published_apphook` import to fail.
